Make sure we are in the EC2 instance.
Added to the yaml file to create a text file in the ec2 instance. Done by creating a new host that is delegated to perform the tasks. Some additional tasks including generating the local_ip of the instance, waiting for the instance to spin up before SSHing, and adding the instance to the host list.